[Post] 로그인 회원가입 이후 구현

+++ backend/src/main/java/com/project/backend/post/service/PostService.java
@@ -6,6 +6,7 @@ import com.project.backend.post.dto.PostListResponseDto;
 import com.project.backend.post.dto.PostResponseDto;
 import com.project.backend.post.dto.PostSaveRequestDto;
 import com.project.backend.post.dto.PostUpdateRequestDto;
+import com.project.backend.security.domain.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,18 +21,23 @@ public class PostService {
     private final PostRepository postRepository;
 
     @Transactional
-    public Long save(PostSaveRequestDto requestDto){
-        return postRepository.save(requestDto.toPostEntity()).getPid();
+    public Long save(PostSaveRequestDto requestDto, User user){
+        return postRepository.save(requestDto.toPostEntity(user)).getPid();
     }
 
     @Transactional
-    public Long update(Long pid, PostUpdateRequestDto requestDto){
+    public String update(Long pid, PostUpdateRequestDto requestDto, User user){
         Post post = postRepository.findById(pid).
                 orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id="+ pid));
+        if((post.getUser()).getUid().equals(user.getUid())){
+            post.update(requestDto.getTitle(), requestDto.getContent(), requestDto.getCategory(), requestDto.getFilename(), requestDto.getFilepath());
+            return "수정이 완료되었습니다.";
+        }
+        else{
+            return post.getUser()+"해당 글쓴이가 아닙니다."+user;
+        }
 
-        post.update(requestDto.getTitle(), requestDto.getContent(), requestDto.getCategory(), requestDto.getFilename(), requestDto.getFilepath());
 
-        return pid;
     }
 
     public PostResponseDto findById(Long pid){
@@ -43,11 +49,18 @@ public class PostService {
     }
 
     @Transactional
-    public void delete(Long pid){
+    public String delete(Long pid, User user){
         Post post = postRepository.findById(pid).
                 orElseThrow(()->new
                         IllegalArgumentException("해당 게시글이 없습니다. id" + pid));
-        postRepository.delete(post);
+        if((post.getUser()).getUid().equals(user.getUid())){
+            postRepository.delete(post);
+            return "삭제가 완료되었습니다.";
+        }
+        else{
+            return "해당 글쓴이가 아닙니다.";
+        }
+
     }
 
     @Transactional(readOnly = true)
@@ -77,6 +90,14 @@ public class PostService {
     public List<PostListResponseDto> searchContent(String keyword){
         return postRepository.findByContentContaining(keyword).stream()
                 .sorted(Comparator.comparing(Post::getPid).reversed())
+                .map(PostListResponseDto::new)
+                .collect(Collectors.toList());
+    }
+
+    @Transactional
+    public List<PostListResponseDto> mypagePost(User user){
+        return postRepository.findByUser(user).stream()
+                .sorted(Comparator.comparing(Post::getPid).reversed()) //먼저 만들어진게 아래 오도록
                 .map(PostListResponseDto::new)
                 .collect(Collectors.toList());
     }
